make json faster for intellisense?
