+++ package.json
@@ -106,7 +106,7 @@
     "react-styleguidist": "^13.0.0",
     "tailwindcss": "^3.0.18",
     "ts-node": "^10.9.1",
-    "typescript": "^4.9.5"
+    "typescript": "^5.0.2"
   },
   "overrides": {
     "@react-pdf/renderer": {
