+++ src/server/trpc/router/courses.ts
@@ -1,26 +1,11 @@
 import { Prisma } from '@prisma/client';
 import { router, publicProcedure } from '../trpc';
 
-import rawCourses from '@data/courses.json';
+import courses, { JSONCourse } from '@data/courses.json';
 
-export type RequisiteType = {
-  type: string;
-  options: JSONCourse[];
-  required: number;
-};
-
-export type JSONCourse = {
-  id: string;
-  co_or_pre_requisites: RequisiteType;
-  corequisites: RequisiteType;
-  course_number: string;
-  prerequisites: RequisiteType;
-  subject_prefix: string;
-  title: string;
-};
 export const coursesRouter = router({
   publicGetAllCourses: publicProcedure.query(async ({ ctx }) => {
-    return rawCourses as unknown as JSONCourse;
+    return courses as JSONCourse[];
   }),
   publicGetSanitizedCourses: publicProcedure.query(async ({ ctx }) => {
     const courses = await ctx.platformPrisma.courses.findMany({
